requirements.txt
tensorflow
gdown
keras
matplotlib
opencv-python
scikit-learn
pandas

fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv